requirements.txt
google-api-python-client==2.172.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
openai==1.86.0
python-dotenv==1.1.0
Flask==3.1.1
requests==2.32.4
jsonify==0.5
flask-cors==6.0.1
gunicorn==23.0.0
tiktoken==0.9.0
google-auth==2.15.0
datetime==5.5